Refactored analytics state management into subdirectory
