+++ src-tauri/src/analytics/mod.rs
@@ -1,8 +1,4 @@
 pub mod algorithms;
 pub mod aux_functions;
-pub mod configuration;
-pub mod controller;
 pub mod data_structures;
-pub mod history;
-pub mod results_store;
 pub mod state;
