+++ src-tauri/src/analytics/state/configuration.rs
