+++ src-tauri/src/analytics/state/history.rs
