+++ src-tauri/src/constructors/mock/mocks.rs
@@ -2,7 +2,9 @@
 
 use crate::analytics::data_structures::neighbor_info::{Neighbor, NeighborInfo};
 use crate::constructors::init::init_edge_map::as_key;
-use crate::data_conversion::distance_conversion::{get_distance, gps_degrees_to_protobuf_field};
+use crate::data_conversion::distance_conversion::{
+    get_spherical_distance, gps_degrees_to_protobuf_field,
+};
 use crate::mesh::device::helpers::get_current_time_u32;
 use crate::mesh::device::MeshNode;
 use app::protobufs;
@@ -139,7 +141,7 @@ pub fn mock_edge_map_from_loc_info(
     for (node_id, node) in nodes.iter() {
         for (neighbor_id, neighbor) in nodes.iter() {
             if node_id != neighbor_id && !edge_map.contains_key(&as_key(*neighbor_id, *node_id)) {
-                let distance = get_distance(Some(node), Some(neighbor));
+                let distance = get_spherical_distance(Some(node), Some(neighbor));
                 if distance < r {
                     let snr = nodes.get(neighbor_id).unwrap().data.snr;
                     let time = get_current_time_u32();
