+++ src-tauri/src/ipc/commands.rs
@@ -2,7 +2,7 @@ use crate::analytics;
 use crate::analytics::algorithms::articulation_point::results::APResult;
 use crate::analytics::algorithms::diffusion_centrality::results::DiffCenResult;
 use crate::analytics::algorithms::stoer_wagner::results::MinCutResult;
-use crate::analytics::configuration::AlgorithmConfigFlags;
+use crate::analytics::state::configuration::AlgorithmConfigFlags;
 use crate::mesh::{self, serial_connection::MeshConnection};
 use crate::state;
 
@@ -286,7 +286,7 @@ pub async fn run_algorithms(
 
     println!("Running algorithms with flags:\n{:#?}", flags);
 
-    state.add_graph(&graph_struct.graph);
+    state.add_graph_snapshot(&graph_struct.graph);
     state.set_algorithm_flags(flags);
     state.run_algos();
     let algo_result = state.get_algo_results();
